# backend/mpay-merchant/src/main/java/com/mapy/merchant/dto/AccountCreateRequest.java
package com.mapy.merchant.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import java.util.Map;
import lombok.Data;

@Data
public class AccountCreateRequest {

    @NotBlank
    private String platform;

    @NotBlank
    private String account;

    @NotBlank
    private String password;

    @NotNull
    private Integer pattern;

    private Map<String, Object> params;
}